Remove old files; Clean up supergraph; Clean up graphframe; Add separate classes for split callee and entry functions.
