+++ callflow/datastructures/ensemblegraph.py
@@ -1,67 +1,85 @@
-import os
+# Copyright 2017-2020 Lawrence Livermore National Security, LLC and other
+# CallFlow Project Developers. See the top-level LICENSE file for details.
+#
+# SPDX-License-Identifier: MIT
+
+#------------------------------------------------------------------------------
+# Library imports
 import networkx as nx
 import pandas as pd
+
+#------------------------------------------------------------------------------
+# CallFlow imports
 import callflow
-from callflow import GraphFrame, SuperGraph
-
 LOGGER = callflow.get_logger(__name__)
+from callflow import SuperGraph
 
-
+# ----------------------------------------------------------------------------
 class EnsembleGraph(SuperGraph):
     """
-    TODO: Clean this up.
-    SuperGraph that handles the ensemble processing.
+    Ensemble SuperGraph Class to handle the processing of ensemble of call graphs.
+
+    Note: I am thinking this might also not really be a class that extends SuperGraph.
     """
 
+    # ------------------------------------------------------------------------
     def __init__(self, props={}, tag="", mode="process", supergraphs={}):
-        # this stores the mapping for each run's data (i.e., Dataset)
+        """
+        Arguments:
+            supergraphs (dict): dictionary of supergraphs keyed by their tag.
+        """
         self.supergraphs = supergraphs
 
         super().__init__(props, tag, mode)
 
-        # For each callsite we store the vector here.
-        self.vector = {}
-
+    # ------------------------------------------------------------------------
     def create_gf(self):
         """
-        Create the graphframes for the ensemble operation.
-        """
-        # Set the gf as first of the dataset's gf
-        if self.mode == "process":
-            first_dataset = list(self.supergraphs.keys())[0]
-            LOGGER.debug(f"Base for the union operation is: {first_dataset}")
+        Create a new callflow.graphframe containing the information of the ensemble.
+        If mode is process, union operation is performed on the df and graph.
+        If mode is render, corresponding files from .callflow/ensemble are read.
 
-            # TODO: do a deep copy.
-            # Instead of a deep copy, create a new graphframe and return it.
-            self.gf = self.supergraphs[first_dataset].gf
+        Note: Code for render is same as in SuperGraph class. Might have to find a way to avoid repetition.
+        """
+        if self.mode == "process":
+
+            self.gf = callflow.GraphFrame()
             self.gf.df = self.union_df()
-            # There is no way to convert networkX to hatchet graph yet. So we are setting this to None.
+            """
+            TODO: Need to write a module to convert a NetowrkX graph to a Hatchet graph.
+            Currently, there is no way to convert networkX to hatchet graph yet. So we are setting this to None.
+            """
             self.gf.graph = None
             self.gf.nxg = self.union_nxg()
 
             assert isinstance(self.gf, callflow.GraphFrame)
 
         elif self.mode == "render":
+            import os
+
             path = os.path.join(self.dirname, self.tag)
 
             self.gf = callflow.GraphFrame()
             self.gf.read(path)
 
-            #parameters = SuperGraph.read_parameters(path)   #TODO: where is this supposed to go?
+            # Read only if "read_parameters" is specified in the config file.
+            if self.props["read_parameter"]:
+                self.parameters = SuperGraph.read_parameters(path)
             self.auxiliary_data = SuperGraph.read_auxiliary_data(path)
 
-            #self.create_gf(data=data)
-            #self.auxiliary_data = self.read_auxiliary_data()
-
+            # NOTE: I dont think we need this anymore. But keeping it just in case.
             # with self.timer.phase(f"Creating the data maps."):
             #     self.cct_df = self.gf.df[self.gf.df["name"].isin(self.gf.nxg.nodes())]
             #     self.create_ensemble_maps()
             #     for dataset in self.props["dataset_names"]:
             #         self.create_target_maps(dataset)
 
+    # --------------------------------------------------------------------
     def union_df(self):
         """
         Union the dataframes.
+        Return:
+            (pd.DataFrame) DataFrame for union of the dataframes.
         """
         df = pd.DataFrame([])
         for idx, tag in enumerate(self.supergraphs):
@@ -75,6 +93,9 @@ class EnsembleGraph(SuperGraph):
     def union_nxg(self):
         """
         Union the netwprkX graph.
+
+        Return:
+            (nx.DiGraph) NetworkX graph for union of graphs.
         """
         nxg = nx.DiGraph()
         for idx, tag in enumerate(self.supergraphs):
@@ -85,16 +106,15 @@ class EnsembleGraph(SuperGraph):
         return nxg
 
     # Return the union of graphs G and H.
-    def union_nxg_recurse(self, nxg_1, nxg_2, name=None, rename=(None, None)):
+    def union_nxg_recurse(self, nxg_1, nxg_2):
         """
-        Iterative concatenation of nodes from nxg_2 to nxg_1.
+        Pairwise Iterative concatenation of nodes from nxg_2 to nxg_1.
+
         """
         if not nxg_1.is_multigraph() == nxg_2.is_multigraph():
             raise nx.NetworkXError("G and H must both be graphs or multigraphs.")
 
         nxg_1.update(nxg_2)
-
-        renamed_nodes = self.add_prefix(nxg_1, rename[1])
 
         is_same = set(nxg_1) == set(nxg_2)
         LOGGER.debug(f"Nodes in Graph 1 and Graph 2 are same? : {is_same}")
@@ -113,44 +133,22 @@ class EnsembleGraph(SuperGraph):
         nxg_1.add_nodes_from(nxg_2)
         nxg_1.add_edges_from(new_edges)
 
-        # # add node attributes for each run
-        # for n in renamed_nodes:
-        #     self.add_node_attributes(nxg_1, n, name)
-
         return nxg_1
 
-    # rename graph to obtain disjoint node labels
-    def add_prefix(self, graph, prefix):
-        if prefix is None:
-            return graph
+    # TODO:
+    def edge_weight(self, nxg):
+        pass
 
-        def label(x):
-            if is_string_like(x):
-                name = prefix + x
-            else:
-                name = prefix + repr(x)
-            return name
-
-        return nx.relabel_nodes(graph, label)
-
+    # TODO
     def add_edge_attributes(self):
-        number_of_runs_mapping = self.number_of_runs()
+        edge_weight_mapping = self.edge_weight()
         nx.set_edge_attributes(
             self.union, name="number_of_runs", values=number_of_runs_mapping
         )
 
-    def number_of_runs(self):
-        ret = {}
-        for idx, name in enumerate(self.unionuns):
-            for edge in self.unionuns[name].edges():
-                if edge not in ret:
-                    ret[edge] = 0
-                ret[edge] += 1
-        return ret
-
-    def add_node_attributes(self, H, node, dataset_name):
+    # TODO
+    def add_node_attributes(self, nxg, node, dataset_name):
         """
-        TODO: Hoist this information to the df directly.
         """
         for idx, (key, val) in enumerate(H.nodes.items()):
             if dataset_name not in self.nxg.nodes[node]:
