+++ callflow/datastructures/supergraph.py
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: MIT
 
-# ------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 # Library imports
 import os
 import json
@@ -11,9 +11,8 @@ import copy
 import pandas as pd
 import numpy as np
 import networkx as nx
-from ast import literal_eval as make_list
 
-# ------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 # CallFlow imports
 import callflow
 from callflow.timer import Timer
@@ -22,77 +21,76 @@ from callflow.modules import EnsembleAuxiliary, SingleAuxiliary
 
 LOGGER = callflow.get_logger(__name__)
 
-# ------------------------------------------------------------------------------
-# SuperGraph Class
+# ----------------------------------------------------------------------------
 class SuperGraph(object):
-
+    """
+    SuperGraph class to handle processing of a an input Dataset.
+    """
     # --------------------------------------------------------------------------
     _FILENAMES = {"params": "env_params.txt", "aux": "auxiliary_data.json"}
 
     # --------------------------------------------------------------------------
-    # --------------------------------------------------------------------------
     def __init__(self, props={}, tag="", mode="process"):
-
+        """
+        Arguments:
+            props (dict): dictionary to store the configuration. CallFlow appends more information while processing.
+            tag (str): Tag for each call graph.
+            mode (str): process|render. process performs pre-processing, and render calculates layout for the client.
+        """
         self.timer = Timer()
 
-        # Props is the information contained in config object.
-        # We duplicate this to add more information to config
-        # and not modify it as a side effect.
         self.props = props
         self.dirname = self.props["save_path"]
-
-        # it appears we're using name as "union", "filter", etc.
-        # this is not a data set name!
         self.tag = tag
-
-        # Mode is either process or render.
         self.mode = mode
-
-        # # Variables used by `create_target_maps`.
-        # self.target_df = {}
-        # self.target_modules = {}
-        # self.target_module_group_df = {}
-        # self.target_module_name_group_df = {}
-        # self.target_module_callsite_map = {}
-        # self.target_module_time_inc_map = {}
-        # self.target_module_time_exc_map = {}
-        # self.target_name_time_inc_map = {}
-        # self.target_name_time_exc_map = {}
-        self.projection_data = {}
 
         self.create_gf()
 
     def create_gf(self):
-        # Create a graphframe based on the mode.
+        """Create a graphframe based on the mode.
+        If mode is process, union operation is performed on the df and graph.
+        If mode is render, corresponding files from .callflow/ensemble are read.
+        """
         if self.mode == "process":
             self.gf = callflow.GraphFrame.from_config(self.props, self.tag)
 
         elif self.mode == "render":
-
             path = os.path.join(self.dirname, self.tag)
 
             self.gf = callflow.GraphFrame()
             self.gf.read(path)
 
-            # parameters = SuperGraph.read_parameters(path)   #TODO: where is this supposed to go?
+            # Read only if "read_parameters" is specified in the config file.
+            if self.props["read_parameter"]:
+                self.parameters = SuperGraph.read_parameters(path)
+
             self.auxiliary_data = SuperGraph.read_auxiliary_data(path)
 
-            # self.create_gf(data=data)
-            # self.auxiliary_data = self.read_auxiliary_data()
-
+            # NOTE: I dont think we need this anymore. But keeping it just in case.
             # with self.timer.phase(f"Creating the data maps."):
             #     self.cct_df = self.gf.df[self.gf.df["name"].isin(self.gf.nxg.nodes())]
             #     self.create_ensemble_maps()
             #     for dataset in self.props["dataset_names"]:
             #         self.create_target_maps(dataset)
 
-    # --------------------------------------------------------------------------
+    # -------------------------------------------------------------------------
+    # Question: Probably belongs to graphframe class?
     def get_module_name(self, callsite):
+        """
+        Get the module name for a callsite.
+        Note: The module names can be specified using the config file.
+        If such a mapping exists, this function returns the module based on mapping. Else, it queries the graphframe for a module name.
+
+        Return:
+            module name (str) - Returns the module name
+        """
         if callsite in self.props["callsite_module_map"]:
             return self.props["callsite_module_map"][callsite]
-        # else:
+
         return self.gf.lookup_with_name(callsite)["module"].unique()[0]
 
+    # -------------------------------------------------------------------------
+    # Remove this block enitrely.
     def _remove__getter(self):
         """
         Getter for graphframe. Returns the graphframe.
@@ -107,24 +105,15 @@ class SuperGraph(object):
 
         self.gf = gf
 
-    # --------------------------------------------------------------------------
-    def _remove_create_gf(self, data=None):
-        """
-        Creates a graphframe using config and networkX graph from hatchet graph.
-        Each graphframe is tagged by a unique identifier.
-        e.g., here is the runName from config file or JSON.
-        """
-        if data:
-            self.gf = callflow.GraphFrame.from_data(data)
-        else:
-            gf = callflow.GraphFrame.from_config(self.props, self.tag)
-            self.gf = copy.deepcopy(gf)
-
-    # --------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
+    # The next block of functions attach the calculated result to the variable `gf`.
     def process_gf(self):
         """
         Process graphframe to add properties depending on the format.
         Current processing is supported for hpctoolkit and caliper.
+
+        Note: Process class follows a builder pattern.
+        (refer: https://en.wikipedia.org/wiki/Builder_pattern#:~:text=The%20builder%20pattern%20is%20a,Gang%20of%20Four%20design%20patterns.)
         """
         if self.props["format"][self.tag] == "hpctoolkit":
 
@@ -158,7 +147,6 @@ class SuperGraph(object):
 
         self.gf = process.gf
 
-    # --------------------------------------------------------------------------
     def group_gf(self, group_by="module"):
         """
         Group the graphframe based on `group_by` parameter.
@@ -176,7 +164,8 @@ class SuperGraph(object):
             filter_perc=self.props["filter_perc"],
         ).gf
 
-    # --------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
+    # Remove this block entirely.
     def _remove_read_gf(self, read_parameter=True, read_graph=False):
         """
         # Read a single dataset stored in .callflow directory.
@@ -229,7 +218,6 @@ class SuperGraph(object):
             df_file_path = os.path.join(dirname, self.tag, df_file_name)
             gf.df.to_csv(df_file_path)
 
-        # TODO: Writing fails.
         if write_nxg:
             nxg_file_name = "nxg.json"
             nxg_file_path = os.path.join(dirname, self.tag, nxg_file_name)
@@ -242,7 +230,8 @@ class SuperGraph(object):
             with open(graph_filepath, "a") as hatchet_graphFile:
                 hatchet_graphFile.write(self.gf.tree(color=False))
 
-    # --------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
+    # Question: These functions just call another class, should we just call the corresponding classes directly?
     def write_gf(self, write_df=True, write_graph=False, write_nxg=True):
         path = os.path.join(self.props["save_path"], self.tag)
         self.gf.write(path, write_df, write_graph, write_nxg)
@@ -257,7 +246,8 @@ class SuperGraph(object):
         SingleAuxiliary(self.gf, dataset=dataset, props=self.props,
                                  MPIBinCount=binCount, process=process)
 
-    # --------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
+    # Read/Write functions for parameter file, auxiliary information (for the client), and pair-wise similarity.
     @staticmethod
     def read_parameters(path):
 
@@ -283,8 +273,6 @@ class SuperGraph(object):
             data = json.load(fptr)
         return data
 
-    # ------------------------------------------------------------------------------
-    # TODO: if this has any future, make this a static function
     @staticmethod
     def _unused_write_similarity(datasets, states, type):
         """
@@ -305,8 +293,9 @@ class SuperGraph(object):
         with open(similarity_filepath, "w") as json_file:
             json.dump(ret, json_file)
 
-    # ----------------------------------------------------------------------------
+    # -------------------------------------------------------------------------
     # NetworkX graph utility functions.
+    # TODO: Remove this in the end.
     def remove_create_target_maps(self, dataset):
         # Reduce the entire_df to respective target dfs.
         self.target_df[dataset] = self.gf.df.loc[self.gf.df["dataset"] == dataset]
@@ -362,312 +351,3 @@ class SuperGraph(object):
         # Exclusive time maps for the module level and callsite level.
         self.module_time_exc_map = self.module_group_df["time"].max().to_dict()
         self.name_time_exc_map = self.module_name_group_df["time"].max().to_dict()
-
-    # -------------------------------------------------------------------------
-
-    # -------------------------------------------------------------------------
-    # Module hierarchy.
-    # TODO: we might have to delete the module hierarchy file in modules later.
-    # TODO: This might fail.
-
-    @staticmethod
-    def _create_source_targets(path):
-        module = ""
-        edges = []
-
-        for idx, callsite in enumerate(path):
-            if idx == len(path) - 1:
-                break
-
-            source = sanitizeName(path[idx])
-            target = sanitizeName(path[idx + 1])
-
-            edges.append({"source": source, "target": target})
-        return edges
-
-    @staticmethod
-    def _check_cycles(hierarchy, G):
-        try:
-            cycles = list(nx.find_cycle(self.hierarchy, orientation="ignore"))
-        except:
-            cycles = []
-
-        return cycles
-
-    @staticmethod
-    def _remove_cycles(hierarchy, G, cycles):
-        for cycle in cycles:
-            source = cycle[0]
-            target = cycle[1]
-            print("Removing edge:", source, target)
-            if source == target:
-                print("here")
-                G.remove_edge(source, target)
-                G.remove_node(source)
-                G.remove_node
-
-            if cycle[2] == "reverse":
-                print("Removing edge:", source, target)
-                G.remove_edge(source, target)
-        return G
-
-    @staticmethod
-    def _add_hierarchy_paths(self, hierarchy, df, path_name, filterTopCallsites=False):
-        module_df = self.df.loc[self.df["module"] == self.module]
-        if filterTopCallsites:
-            group_df = module_df.groupby(["name"]).mean()
-            f_group_df = group_df.loc[group_df[self.config.filter_by] > 500000]
-            callsites = f_group_df.index.values.tolist()
-            df = df[df["name"].isin(callsites)]
-
-        paths = df[path_name].unique()
-        for idx, path in enumerate(paths):
-            if isinstance(path, float):
-                return []
-            path = make_tuple(path)
-            source_targets = self.create_source_targets(path)
-            for edge in source_targets:
-                source = edge["source"]
-                target = edge["target"]
-                if not hierarchy.has_edge(source, target):
-                    hierarchy.add_edge(source, target)
-
-    def module_hierarchy(self, module=None):
-        hierarchy = nx.DiGraph()
-        node_paths_df = self.df.loc[self.df["module"] == self.module]
-
-        if "component_path" not in self.df.columns:
-            utils.debug("Error: Component path not defined in the df")
-
-        with self.timer.phase("Add paths"):
-            self._add_hierarchy_paths(hierarchy, node_paths_df, "component_path")
-
-        cycles = self._check_cycles(hierarchy)
-        while len(cycles) != 0:
-            self.hierarchy = self._remove_cycles(hierarchy, cycles)
-            cycles = self._check_cycles(hierarchy)
-            print(f"cycles: {cycles}")
-
-        return hierarchy
-
-    # ------------------------------------------------------------------------------
-    # Reveal a callsite's path
-    # TODO: not tested. Could break.
-    def create_source_targets(self, component_path):
-        module = ""
-        edges = []
-        for idx, callsite in enumerate(component_path):
-            if idx == 0:
-                module = component_path[0]
-                edges.append(
-                    {
-                        "module": module,
-                        "source": module,
-                        "target": module + "=" + component_path[idx + 1],
-                    }
-                )
-                pass
-            elif idx == len(component_path) - 1:
-                pass
-            else:
-                edges.append(
-                    {
-                        "module": module,
-                        "source": module + "=" + component_path[idx],
-                        "target": module + "=" + component_path[idx + 1],
-                    }
-                )
-
-        return edges
-
-    def callsite_paths(self, callsites):
-        paths = []
-        for callsite in callsites:
-            df = self.name_group_df.get_group(callsite)
-            paths.append(
-                {
-                    "group_path": make_list(df["group_path"].unique()[0]),
-                    "path": make_list(df["path"].unique()[0]),
-                    "component_path": make_list(df["component_path"].unique()[0]),
-                }
-            )
-        return paths
-
-    def add_reveal_paths(self, reveal_callsites):
-        # TODO: this will fail because there is no self.supergraph
-        assert False
-        paths = self.callsite_paths(reveal_callsites)
-
-        for path in paths:
-            component_edges = self.create_source_targets(path["component_path"])
-            for idx, edge in enumerate(component_edges):
-                module = edge["module"]
-
-                # format module +  '=' + callsite
-                source = edge["source"]
-                target = edge["target"]
-
-                if not self.gf.nxg.has_edge(source, target):
-                    if idx == 0:
-                        source_callsite = source
-                        source_df = self.module_group_df.get_group((module))
-                        source_node_type = "super-node"
-                    else:
-                        source_callsite = source.split("=")[1]
-                        source_df = self.module_name_group_df.get_group(
-                            (module, source_callsite)
-                        )
-                        source_node_type = "component-node"
-
-                    target_callsite = target.split("=")[1]
-                    target_df = self.module_name_group_df.get_group(
-                        (module, target_callsite)
-                    )
-                    target_node_type = "component-node"
-
-                    source_weight = source_df["time (inc)"].max()
-                    target_weight = target_df["time (inc)"].max()
-
-                    edge_type = "normal"
-
-                    print(f"Adding edge: {source_callsite}, {target_callsite}")
-                    self.gf.nxg.add_node(source, attr_dict={"type": source_node_type})
-                    self.gf.nxg.add_node(target, attr_dict={"type": target_node_type})
-                    self.gf.nxg.add_edge(
-                        source,
-                        target,
-                        attr_dict=[
-                            {
-                                "source_callsite": source_callsite,
-                                "target_callsite": target_callsite,
-                                "edge_type": edge_type,
-                                "weight": target_weight,
-                                "edge_type": "reveal_edge",
-                            }
-                        ],
-                    )
-
-    def add_exit_callsite():
-        assert False
-        # TODO: This code is missing for some reason.
-        pass
-
-    # --------------------------------------------------------------------------
-    # Create a module hierarchy for a chosen module.
-    # Not fully tested. Might break.
-    # TODO: these functions should be private (preprend _) and static
-    def module_entry_functions_map(self, graph):
-        entry_functions = {}
-        for edge in graph.edges(data=True):
-            attr_dict = edge[2]["attr_dict"]
-            edge_tuple = (edge[0], edge[1])
-            for edge_attr in attr_dict:
-                if edge_tuple[1] not in entry_functions:
-                    entry_functions[edge_tuple[1]] = []
-                entry_functions[edge_tuple[1]].append(edge_attr["target_callsite"])
-        return entry_functions
-
-    def create_source_targets_from_group_path(self, path):
-        module = ""
-        edges = []
-        for idx, callsite in enumerate(path):
-            if idx == len(path) - 1:
-                break
-            source = path[idx].split("=")
-            target = path[idx + 1].split("=")
-            edges.append(
-                {
-                    "source": source[0],
-                    "target": target[0],
-                    "source_callsite": source[1],
-                    "target_callsite": target[1],
-                }
-            )
-        return edges
-
-    def same_source_edges(self, component_edges, reveal_module):
-        ret = []
-        for idx, edge in enumerate(component_edges):
-            source = edge["source"]
-            target = edge["target"]
-
-            if source == reveal_module:
-                ret.append(edge)
-        return ret
-
-    def same_target_edges(self, component_edges, reveal_module):
-        ret = []
-        for idx, edge in enumerate(component_edges):
-            source = edge["source"]
-            target = edge["target"]
-
-            if target == reveal_module:
-                ret.append(edge)
-        return ret
-
-    def add_entry_callsite(self, reveal_module):
-        entry_functions_map = self.module_entry_functions_map(self.gf.nxg)
-        reveal_callsites = entry_functions_map[reveal_module]
-        paths = self.callsitePathInformation(reveal_callsites)
-
-        for path in paths:
-            component_edges = self.create_source_targets_from_group_path(
-                path["group_path"]
-            )
-            source_edges_to_remove = self.same_source_edges(
-                component_edges, reveal_module
-            )
-            target_edges_to_remove = self.same_target_edges(
-                component_edges, reveal_module
-            )
-
-            if len(source_edges_to_remove) != 0:
-                for edge in source_edges_to_remove:
-                    if self.gf.nxg.has_edge(edge["source"], edge["target"]):
-                        self.gf.nxg.remove_edge((edge["source"], edge["target"]))
-                    self.gf.nxg.add_node(
-                        reveal_module + "=" + edge["source_callsite"],
-                        attr_dict={"type": "component-node"},
-                    )
-                    self.gf.nxg.add_edge(
-                        (reveal_module + "=" + edge["source_callsite"], edge["target"]),
-                        attr_dict=[
-                            {
-                                "source_callsite": edge["source_callsite"],
-                                "target_callsite": edge["target_callsite"],
-                                "edge_type": "normal",
-                                "weight": self.module_name_group_df.get_group(
-                                    (reveal_module, edge["source_callsite"])
-                                )["time (inc)"].max(),
-                                "edge_type": "reveal_edge",
-                            }
-                        ],
-                    )
-
-            if len(target_edges_to_remove) != 0:
-                for edge in target_edges_to_remove:
-                    if self.gf.nxg.has_edge(edge["source"], edge["target"]):
-                        self.gf.nxg.remove_edge(edge["source"], edge["target"])
-                    self.gf.nxg.add_node(
-                        reveal_module + "=" + edge["target_callsite"],
-                        attr_dict={"type": "component-node"},
-                    )
-                    self.gf.nxg.add_edge(
-                        edge["source"],
-                        reveal_module + "=" + edge["target_callsite"],
-                        attr_dict=[
-                            {
-                                "source_callsite": edge["source_callsite"],
-                                "target_callsite": edge["target_callsite"],
-                                "edge_type": "normal",
-                                "weight": self.module_name_group_df.get_group(
-                                    (edge["target"], edge["target_callsite"])
-                                )["time (inc)"].max(),
-                                "edge_type": "reveal_edge",
-                            }
-                        ],
-                    )
-
-        self.gf.nxg.remove_node(reveal_module)
-
-    # --------------------------------------------------------------------------
