+++ callflow/modules/module_hierarchy.py
@@ -1,48 +1,40 @@
-import math
-import time
+# Copyright 2017-2020 Lawrence Livermore National Security, LLC and other
+# CallFlow Project Developers. See the top-level LICENSE file for details.
+#
+# SPDX-License-Identifier: MIT
 
-import pandas as pd
+
+# ------------------------------------------------------------------------------
+# Library imports
 import networkx as nx
-from networkx.readwrite import json_graph
 
-from ast import literal_eval as make_tuple
-
+# ------------------------------------------------------------------------------
+# CallFlow imports
 import callflow
-
 LOGGER = callflow.get_logger(__name__)
-from callflow.timer import Timer
-
 
 class ModuleHierarchy:
-    def __init__(self, supergraph, module):
+    def __init__(self, supergraph, module, filter_by="time (inc)", filter_perc=10.0):
         self.df = supergraph.gf.df
         self.module = module
 
-        # Create the Super node's hierarchy.
-        self.hierarchy = nx.DiGraph()
-        self.timer = Timer()
+        self.hierarchy = ModuleHierarchy.create_nxg_tree_from_paths(df=supergraph.gf.df, path="path", filter_by=filter_by, filter_perc=filter_perc )
 
-        self.result = self.run()
+        cycles = self.check_cycles(self.hierarchy)
+        while len(cycles) != 0:
+            self.hierarchy = self.remove_cycles(self.hierarchy, cycles)
+            cycles = self.check_cycles(self.hierarchy)
+            print(f"cycles: {cycles}")
 
-    def create_source_targets(self, path):
-        module = ""
-        edges = []
-
-        for idx, callsite in enumerate(path):
-            if idx == len(path) - 1:
-                break
-
-            source = callflow.utils.sanitize_name(path[idx])
-            target = callflow.utils.sanitize_name(path[idx + 1])
-
-            edges.append({"source": source, "target": target})
-        return edges
-
-    def add_paths(self, df, path_name, filterTopCallsites=False):
-        module_df = self.df.loc[self.df["module"] == self.module]
-        if filterTopCallsites:
+    @staticmethod
+    def create_nxg_tree_from_paths(self, df, path_name, filter_by, filter_perc):
+        """
+        Create a networkx graph for the module hierarchy.
+        Filter if filter percentage is greater than 0.
+        """
+        if filter_perc > 0.0:
             group_df = module_df.groupby(["name"]).mean()
-            f_group_df = group_df.loc[group_df[self.config.filter_by] > 500000]
+            f_group_df = group_df.loc[group_df[filter_by] > filter_perc * group_df[filter_by].max()]
             callsites = f_group_df.index.values.tolist()
             df = df[df["name"].isin(callsites)]
 
@@ -51,14 +43,38 @@ class ModuleHierarchy:
             if isinstance(path, float):
                 return []
             path = make_tuple(path)
-            source_targets = self.create_source_targets(path)
+            source_targets = ModuleHierarchy._create_source_targets(path)
             for edge in source_targets:
                 source = edge["source"]
                 target = edge["target"]
                 if not self.hierarchy.has_edge(source, target):
                     self.hierarchy.add_edge(source, target)
 
-    def as_spanning_trees(self, G):
+    @staticmethod
+    def _create_source_targets(path):
+        """
+        Create edges from path.
+
+        Params:
+            path (list) - paths expressed as a list.
+
+        Return:
+            edges (array) - edges expressed as source-target pairs.
+        """
+        edges = []
+
+        for idx, callsite in enumerate(path):
+            if idx == len(path) - 1:
+                break
+
+            source = sanitizeName(path[idx])
+            target = sanitizeName(path[idx + 1])
+
+            edges.append({"source": source, "target": target})
+        return edges
+
+    @staticmethod
+    def as_spanning_trees(G):
         """
         For a given graph with multiple sub graphs, find the components
         and draw a spanning tree.
@@ -75,7 +91,14 @@ class ModuleHierarchy:
 
         return ret
 
-    def check_cycles(self, G):
+    @staticmethod
+    def check_cycles(G):
+        """
+        Checks if there are cycles.
+
+        Return:
+            The cycles in the graph.
+        """
         try:
             cycles = list(nx.find_cycle(self.hierarchy, orientation="ignore"))
         except:
@@ -83,7 +106,12 @@ class ModuleHierarchy:
 
         return cycles
 
-    def remove_cycles(self, G, cycles):
+    @staticmethod
+    def _remove_cycles(G, cycles):
+        """
+        Removes cycles from the networkX Graph.
+        TODO: Improve the logic here.
+        """
         for cycle in cycles:
             source = cycle[0]
             target = cycle[1]
@@ -98,24 +126,3 @@ class ModuleHierarchy:
                 print("Removing edge:", source, target)
                 G.remove_edge(source, target)
         return G
-
-    # instead of nid, get by module. nid seems very vulnerable rn.
-    def run(self):
-        node_paths_df = self.df.loc[self.df["module"] == self.module]
-
-        if "component_path" not in self.df.columns:
-            utils.debug("Error: Component path not defined in the df")
-
-        with self.timer.phase("Add paths"):
-            self.add_paths(node_paths_df, "component_path")
-
-        cycles = self.check_cycles(self.hierarchy)
-        while len(cycles) != 0:
-            self.hierarchy = self.remove_cycles(self.hierarchy, cycles)
-            cycles = self.check_cycles(self.hierarchy)
-            print(f"cycles: {cycles}")
-
-        print(len(self.hierarchy.nodes()), len(self.hierarchy.edges()))
-        for edge in self.hierarchy.edges():
-            print(edge)
-        return self.hierarchy
